similarity/requirements.txt
pdftotext==2.1.2
python-docx==0.8.10
